Comments - Defect Number: 3231; Defect RecordID: 3227; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Who has access to the PM Demo environment?; Defect Description: Question: Which users currently have access to the Demo environment? What is the process for giving access?
I have heard some talk (just ideas for now, not solidified for now) from our sales team that they may want to show Practice Managment to an external client as a showcase opportunity. I'm trying to understand what the process and timeframe would be for updating the Demo environment with the current functionality in Production, run the obfuscation script, and give access to the appropriate user so that I can set their expectations for access.

[Harshitha Tenneti]
Demo Site can be accessed with URL: http://65.52.16.240/Dashboard.aspx.

There are 60 active people in the current demo system. Attached excel document contains the people who have access to the system.

Lisa, Currently you will be able access to the site. We are resetting your password to Demo site as the password stored in Demo DB is very old, you will receive an email regarding password reset.

The Demo system was last updated on 2013/03/04. Please let us know if we want to update the Demo Environment with the current Production version.

Lisa Neel:

Thank you so much for this explanation. Is giving access to the Demo environment as simple as adding a new person record to the Demo Environment and then giving them the Demo environment URL?

[Harshitha Tenneti] : Yes Lisa, adding a person to the demo system gives him acess to the system.

I think it's time for another update, just so that we have to use the Demo environment at the last minute, it won't be too out of sync with current functionality. Could you please update the environment and perform the necessary obfuscation scripts?

[Harshitha Tenneti] : Could you please let us know when you want us to update the demo site?