Comments - Defect Number: 3236; Defect RecordID: 3232; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Persons > First and Last Name Fields - allow use of a space (in cases where a name is two words; Defect Description: We are encountering situations where new employees have two last names or two parts of their last name. (Example: "Van Huzien".) Can we please open the rules for the "First Name" and "Last Name" fields to allow a single space in the field?
The error message when entering non-allowed characters should be updated to say: "First/Last Name should be limited to 2-35 characters in length containing only letters and/or an apostrophe, hyphen, or a single space."

Please let me know if there are any unintended consequences as a result of this request. Otherwise, given the small nature of this change, please feel free to push directly to Production.

Thank you very much!


[Harshitha Tenneti]

We have one more field similar to first name and Last name in the system i.e., Buyer name in Project detail and opportunity detail.
Could you please let us know if you want us to work on it also?

LN: Yes please!