Comments - Defect Number: 3242; Defect RecordID: 3238; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Hide "Opportunity" functionality for all Roles/Seniorities except "Administrator"; Defect Description: With the company's transition to Dynamics CRM to track and manage its sales opportunities, we are no longer using any of the functionality under the "Opportunity" heading. I don't want to deprecate the functionality, but I would like to clean up the view for users to make for a more streamlined experience. To that end, please hide the "Opportunity" heading and all functionality underneath for all user roles/seniorities except for those with Administrator privileges.
Please let me know if any questions/issues resulting from this request!

Thanks very much!

[Harshitha Tenneti]:
1. We have a field in Project detail called Opportunity Linking. Do we need to show that field only to Administrator from now on?

LN: Yes please.

2. Consulting demand report: Graphs tab: We show sales stages related to opportunities along with project sales stages. Do you want us to hide opportunity sales stages for persons other than administrator? LN: Thank you for raising this point; the purpose of the Consulting Demand Report was to help the Recruiting team understand hiring needs based on the Opportunities we had in process. Since we no longer use Opportunities, this report is of little use to the Recruiting team. Please hide the report to everyone except Admins.
3. Consulting demand report: Details tab: We show opportunity numbers in details tab. clicking on the number will navigate to opportunity detail page. Do you want us to mark opportunity number as just label instead of link for users other than Administrator? LN: As described above, let's just hide the report in totality.
4. Can you please let us know what should be the action made in following scenario:
a.Consider that a person is added as owner to an existing opportunity.
b.When we try to terminate that person we see a message as ABC,XXX is designated as the Owner for the following Opportunities: {Opportunities list}

LN: Could you please remind me of the rest of the workflow after the validation message? Does the system disallow a person from being terminated if they have open Opportunities, or does the message just serve as a reminder that the person has open Opps? If the message is just a reminder, then no changes to functionality are needed.

[Harshitha Tenneti]:
The message that is displayed in the above scenario is just a reminder to the user.Clicking on 'OK' of the message will terminate that person. Opportunity owner's field in opportunity summary will remain populated with his name.