Comments - Defect Number: 3230; Defect RecordID: 3226; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Milestone Person detail: User is able to add already existing resource for same time period.; Defect Description: From Milestone detail page, user cannot add a project resource with more than one bill rate on a project during the same time period.
A validation message will be displayed to the user as below when he tries to do so:
A project resource cannot have more than one bill rate on a project during the same time period. Please adjust the start and end dates to make sure that the time periods for the bill rates do not overlap.
But user is able to add a resource twice to project with same/different bill rates from milestone person detail page.

Please take a look at the below image:

Steps to reproduce:
1. Add a person as resource in a project milestone
2. Try to add already existing person for same time period.--> A validation message will be displayed to the user
3. Now navigate to milestone person detail page of the resource
4. Click on Add another row to add a new record for the resource
5. Add details for new record with same dates and same/different compensation details.No Validation message will be displayed and the person will be saved

Could you please let us know your reponse on this?