Comments - Defect Number: 3248; Defect RecordID: 3244; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Time > Calendar > PTO Report Export; Defect Description: Right now, we have the ability in the system to enter time off to peoples calendars. We have the ability to visually view planned time off for the next three months via the Utilization Report, but currently there is no way to export those time off results out of Practice Management for reporting purposes.
The following request is the first step for the creation of a new set of reports that will allow us to view time off that has been entered into the system. This includes both time off entered via the Calendar as well as time off that has been entered by the user in Time Entry.

This first step is to develop an Excel export that can be downloaded and then analyzed outside of the system. (After this functionality has been implemented, then Ill look to draft requirements for the development of user-changeable reports in the Practice Management UI. As discussed on our last call, I woudl appreciate your suggestions as well! I imagine a report similiar to the "New Hire" or "Termination" Reports where users can filter by Practice Area, Project Account, etc. in both tabular as well as graphical formats. The user of these reports would be a Project Manager, Business Unit Owner, or member of the Senior Leadership Team who is trying to understand who on their projects or in the company at large is going to be Out of the Office and when so that they can plan for back-up on projects.)

See a mock-up of the small changes needed to the current Calendar UI functionality:

The same security roles that currently have access to the Add Time Off functionality in PM should also be able to download this Excel report.

The export date range will be based on the calendar year showing on the Calendar in the UI (i.e., downloading the Excel file when the 2014 calendar is on screen will pull the report for the period of 1/1/2014  12/31/2014).

The Include Company Holidays checkbox is included for the user to determine whether he/she would like to see the report including company holidays. (The choice is given as the results will be much longer). The default is that the box is unchecked and holiday days are not included the results.

The Excel Export contains the following columns in the following order. Only Pay Type W2-Salary individuals with Active status in the system at that point in time when the report is run are included in the report.
Person ID
Person Name
Time Off Type: PTO, ORT, BER for Bereavement (unless there is another acronym already in use by the system; if so, then use that), HOL. If the Include Company Holidays box is checked, then holiday days are included in the results.
Time Off Start Date
Time Off End Date
Project Number: For the project that the individual is assigned to during that Time Off Start/End Date period. (See description in bullet below.)
Project Name: If the individual is assigned to a project (all Project Status types) within the period of the Time Off Start and End dates.
o Note  if a person is assigned to multiple projects within the same Time Off Start/End period, then I would expect to see two line items returned for that individual. The Time Off Type, Start and End Dates would be the same, but the Project Number and Project Names would be different.
Project Status: Project Status of the project(s) individual is assigned to during the Time Off Start/End period.
Account: Client Account Name of the project(s) individual is assigned to during the Time Off Start/End period.
Business Group: Business Group of the project(s) individual is assigned to during the Time Off Start/End period.
Business Unit: Business Unit of the project(s) individual is assigned to during the Time Off Start/End period.
Practice Area: Practice Area of the project(s) individual is assigned to during the Time Off Start/End period.
Project Manager(s): Project Managers of the projec