Comments - Defect Number: 3253; Defect RecordID: 3249; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: List of people who have less than ten fields selected/completed in the Skills Profile; Defect Description: We have been trying to encourage usage of the Skills Profile amongst the Consultant workforce, but are having a difficult time monitoring and managing completion. Reporting capability is definitely an area that we will need to investigate for this piece of functionality, but for right now, could you please help me with a simple reporting request:
Could you please determine which Active individuals with any Pay Types in any Practice Area EXCEPT Administration and Offshore do NOT have more than ten fields completed in their individual Skills Profile? Woudl you also include the last (latest) date that their Skills Profile was updated?

The report does not have to be substantial - at the minimum, I need the names of the individuals that do not have more than ten fields completed in the Skills Profile and the date of the last time their profile was updated. Feel free to include any other fields in the report that you feel would be useful for his analysis. I plan to use the report to a 1) gauge compliance to the completion of the profile as well as 2) chase individuals who have not completed it.

Thanks very much!