Comments - Defect Number: 3240; Defect RecordID: 3236; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: UAT Email Functionality; Defect Description: Could you please turn off the email from UAT now that we have concluded testing that functionality for the time being? Is it possible to build a script so that even when UAT is updated with a copy of Production, the email address recipient for all emails is me and me alone? This way, even if we forget to turn off UAT email functionality after a round of testing, the UAT emails will still only be sent to me. Other ideas are very welcome!
[Harshitha Tenneti]

We are thinking to modify code in such a way that to update the receiver address as "lisan@logic2020.com" when ever mails are enabled in UAT.

We think that it is better than running a script when ever UAT is updated with a copy of Production, because there may result in chances of forgetting the script to be run in UAT.

Please let us know if you have any concerns on it so that we can think of some better option. We request you to add any ideas you have on this.