Comments - Defect Number: 3260; Defect RecordID: 3256; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Project Summary Page > Ellen Boyer cannot access; Defect Description: I have an unusual issue that is apparently affecting only Ellen. She tries to access the Project Summary page, only to receive a "Contact Administrator" error message.
Reproduction Steps:
1. Log into Practice Management
2. Click Projects > Summary
3. Page returns with "Error Occurred. Please contact your system administrator."


The unusual part of this error is that I cannot replicate it either logged in as mysel for via the "Troubleshooting" functionality logged in as Ellen. I have, however, watched Ellen and can confirm that the issue is real. Could it be a caching issue?

The request above is a hotfix request from a prioritization perspective. Please implement any fix directly to PM.
Harshitha: I think its a browser caching issue.
We Investigated by changing the password of Elen Boyer in back up of Production database.
We see no issues with Project summary page when we login as Elen Boyer from login page.
Could you mention her to clear the browser cache and see the page.
If she still see the issue let us know if there are any other reproduction steps..

The second part of this request is an "Urgent" priority: I have noticed that the Summary Page has difficulty toggling between the "Show Projects for" time filters. I suspect that the automatic update upon changing from one time filter to another (i.e., I've had particular difficulty with Previous FY to Current FY) does not always activate, but I haven't been able to do any more substantial testing. Could you please investigate for me and help me to identify any causes and your recommended fixes? One idea I had is that insert an "Update" button in the main blue bar at the top of the Project Summary Screen (next to the current "Add Project"). Current the "Update" button is hidden in the filters, and not too many know that it exists.
Harshitha: We will move the 'Update' button to the besidesof Add Project button based on your suggestion and will remove automatic update functionality. The summary report will be updated only when we click Update button.
We will update it to UAT tomorrow.

Thanks very much for your help!