Comments - Defect Number: 3265; Defect RecordID: 3261; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Investigation: deleting projects in "Experimental" and "Inactive" status; Defect Description: I'd like to investigate the feasibility of deleting projects with statuses of "Inactive" and "Experimental". There a few business resons driving this request: We don't really use the "Inactive" status as a status indicator any longer. Our sales process has evolved so that projects follow the status of Projected > Active > Completed. Projects that aren't officially sold are moved back into Experimental rather than Inactive. As a result, the "Experimental" status turns into a large "sandbox" with no way of us to clean up items that are not in use.
As a first step, could you please run a report for me of all of the projects in the system that currently have a status of "Experimental" and "Inactive", and the dates that they were last modified? In addition, would you pay particular consideration to projects that were originally "Active" status and were changed to "Inactive"? I need to understand what projects are currently those statuses so that I can see how recent they are (and therefore how pertinent the information continues to be to our company).

Would you please investigate what the system impacts would be if we allowed users to delete a projects that have a status of "Inactive" or "Experimental". My intention is that anyone with access to Project Detail could delete an Experimental project, but deleting an Inactive project would be limited to Senior Leadership Team/Administrators. I am imagining that the user would have two ways to delete a project: when the project is displayed as a line item, like in Project Summary or when searched on, a little "x" icon (used in other places of the system where delete capability is desired) to the right of the line items would appear. The user would click that "x" and then a validation message would confirm the user wants to delete the project. The message should also warn that the deletion cannot be undone. The second way would be a new icon in Project Details at the bottom of the page entitled "Delete Project". The same validation message would appear when that button is clicked.