Comments - Defect Number: 3257; Defect RecordID: 3253; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Analysis and Recommendations Project - Page Load Optimization; Defect Description: One of the key pieces of feedback I hear from Practice Management time and again is how slow the system is to take load pages. The concerns are across the entire system, but the specific pain point for the Consulting organization is the length of time it takes to tab between tabs in Project Details and the length of time it takes to load and add resource details (i.e., adding a new resource to Milestone Details, changing start and end date, etc.)
I understand that we are limited by the system's inherent architecture in many ways, but I would like to get your help in analyzing how we could improve page load and field load times and make the system faster overall. I'm hoping that there are some options available to us to alleviate some of the frustration while not having to get to the point of a complete rearchitecture of the system!

My high-level thinking would be that your analysis would be delivered in three parts:
Analysis of the areas of the system that are slowest to respond and the primary reason for that slowness (I can provide more anecdotal evidence if you'd like!)
Recommended short-, medium-, and long-term fixes, and
A level of difficulty and level of effort it would take to complete each recommendation.

From this analysis, I'll review with leadership and prioritize from there.

Do let me know your thoughts if this would be a valuable exploration!