Comments - Defect Number: 3271; Defect RecordID: 3267; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: New Functionality: Project Feedback Tracking; Defect Description: These are requirements for a new and very important piece of functionality for Practice Management. I would like to prioritize your time right now to review my written requirements and mock-ups attached, and then ask questions concerning any questions/concerns you have about the intended functionality and any gaps in the business logic/scenarios as I've described it here. Ellen may have a few other requests coming down the pipeline next week. We may end up prioritizing development of those items before the actual development of the Project Feedback functionality, but in the meantime I would like to ensure that you have the direction and clarity you need so that you can begin work on this functionality as soon as we give the go ahead!
To give you the business background for this new functionality: Giving our consultants regular, clear, and concrete feedback is a very important concept for our company because we want to encourage people's growth in their career. However, as a company, we have not been good at formally documenting that feedback, and ensuring that every individual has had formal reviews on their projects consistently throughout the year. To help to alleviate that pain in our organization, I wish to implement new functionality that will help us track when Project Feedback forms are due to be completed and track the completion of those Project Feedback forms. Adding this level of automated visibility and tracking is the first step. The leadership team is currently considering what tool in which to actually input the feedback. Right now, our process is that a Word document is sent to an email mailbox and then manually stored. v2 functionality of Project Feedback Tracking may be to bring the current Word form inside of PM so that individuals can input feedback directly into the system. That component is currently not finalized - the scope of the work for this initial v1 effort is just the tracking and administration components I describe below - but I wanted to give you a sense of where we may be going with this functionality in the future.

Project Feedback forms are due for every salaried consultant based on the following policy:
At the end of their project engagement if the engagement is longer than 6 weeks in duration, or
Every 3 months if they are staffed on a project lasting longer than three months in duration.

Based on this policy above, the company's expectation is that consultants have a total of at least four completed project feedback forms per year.

Exceptions to this policy:
Only W2-Salary employees are eligible for documented, formalized project feedback. For this new functionality, do not consider inivduals with a pay type of "W2-Hourly" or "1099".
Individuals with the title of "Client Director" or "Practice Director" are not considered for Project Feedback. Even if assigned to the project as a billing resource, do not consider them in the functionality described below.
Project feedback forms are not required for individuals staffed on projects for less than 6 weeks in duration


There are three components to the new functionality:
1. New tab entitled "Project Feedback" in Project Details
2. New report entitled "Project Feedback"
3. New automated emails to serve as notificaton of action required.

Our timing for this functionality above is to go-live by the first of July. But we're only going to enforce compliance to the policy described above for Review Period Periods that end after July 1, 2014. That means we don't want email reminders sent out for overdue Project Feedback notifications for Project Review Periods that ended prior to July 1st.

1. Project Feedback Tab
The new tab in Project Details tracks the number of reviews due for individuals on the project. Because of the policy above, it is expected that many individuals staffed on longer term engagements will have multiple l