Comments - Defect Number: 3252; Defect RecordID: 3248; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Business Unit Manger is able to add/delete resources to milestone but cannot edit the records.; Defect Description: A person with business unit manager role is able to add/delete resources to milestones of projects that he has permissions but he is unable to edit those resources unless he is the owner/salesperson/project manager of the project.
Steps to reproduce:
Select a person who have 'Business unit manager' role and permissions to all projects. (We changed 'Cheruku, Sainath' role as 'Business unit manager' and set permissions to all projects in UAT)
Go to 'Become user ' page and login as 'Cheruku, Sainath'.
Now open P000370 project (in which Cheruku,sainath is the owner of the project) from 'project search' page.
Click on 'Project' milestone and add any resource in 'Resources' tab.
Resource will be added to the milestone.
You will be able to edit the record,if you try to edit.
You will be able to delete the record,if you try to delete.
Open P000254 (in which 'Cheruku,Sainath' is not owner/salesperson/project manager for the project) project from 'Project Search' page.
Click on 'Assessment' milestone and add any resource in 'Resources' tab.
Resource will be added to the milestone.Try to edit the resource--you will not be able to edit.
Resource will be deleted,if you try to delete the record.

Could you please let us know if it is the expected scenario?