Comments - Defect Number: 3298; Defect RecordID: 3294; Server Address: localhost; Server Port: 99; Database Name: Practice Management; Defect Summary: Include a new drop down in Project summary page to allow user select different filters for the report for the financial calcul@; Defect Description: Project summary page should be added with a new filter for selecting options for financial calculations in the page. The drop down should be added by removing the check box 'Use Actual Time Entry data for Previous Month calculations' from its place.
The newly added drop down should consists of following values in it.
Projected values for all months
Actual values for all months
Actual values for previous months and Projected values for current/Future months.


Default value selected should be 'Projected values for all months'. The export in the page should reflect the same data which is present in the page.

The drop down implementation should look like the attachment.