Comments - Defect Number: 3338; Defect RecordID: 3334; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: Able to change person status from Active to Contingent.; Defect Description: Harshithat,
Would it be possible to change the Status indicator to allow placing a user back into Contingent status as long as they have not charged time? Let me know if that is easy or more complex.

-Nick

Hi Nick,

Please find the following question on changing person status from Active to Contingent who has no time entries entered,
1.We will display a date field when we change person status from contingent to active, do you want us to display a similar date field when we change status from Active to contingent?

Yes, a date field.

You can go ahead and work on this in your free time (until I get the zero bill rate spec complete).

-Nick

Hi Nick,

While working on "Changing person status Active to Contingent" task, we came across following scenario. Please take a look and let us know your response on this.
Consider a person with 'Active' status.
The person is assigned to a project but he does not have any time entries entered, then we are able to change the person status to 'Contingent'.
If we change the person status to 'Contingent' and change "Hire date" so that the "new Hire date" is in between Project from and to dates, then the person dates in project are getting updated according to new hire date even though his status is contingent.
For example, If a person with active status and hire date as 09/01/2015 and he is assigned to a project from 10/01/2015 to 10/30/2015. But he does not have any time entries entered. Now if we change his status to contingent and hire date as 10/15/2015, then the project dates are updated as 10/15/2015 to 10/30/2015 in Person Detail > Projects tab.

But the conditions to populate people inMilestone Detail > Resource tab are
For persons,who have Active status for at least one day during the range of milestone Start date and End date.
Persons with 'Active' or 'Termination Pending' status and with role as 'Client Director' or 'Practice Area Manager'.
In this scenario, Contingent people are assigned to project without considering the condition. To handle this scenario we are thinking the following two options. Could you please suggest us an option to handle the above scenario.
We should either show a warning message saying "Changing Active to contingent will remove projects assigned to the person. Click ok to continue or click cancel to exit without saving" and when user clicks 'Ok' button then we should delete the person from all the projects instead of updating person dates according to recent hire date.
We should not allow user to change person status to contingent if the person is assigned to any project.
@In those instances where a user is assigned to projects already, an error message should be presented: This user is currently assigned to one or more projects. Please remove these resource allocations before changing to Contingent status.

-Nick