Comments - Defect Number: 3340; Defect RecordID: 3336; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: Logo changes task; Defect Description: Good morning!
I am trying tochangethelogoin PMwed like thelogoto NOT have that broad white bandthe attached is a transparent backgroundlogothat I tried to upload, but it is still showing as a white backgroundcan you put it in place smoothed into the blue background of the applications (like the priorlogo)?

Thanks!

-Nick

Hi Nick,

Please find the following points and let us know your response.
We have deployedlogochangesto UAT. Please verify and let us know if you want us to deploy thesechangesto Production.
We have modified the footerlogosent by you for proper alignment.
Thelogoof favicon(shown in below image) is notchanged. Could you please let us know andprovide the imageif you want us tochangethatlogoalso.
We are thinking of adding a preview for the uploaded image in "Branding/Logo"(Configuration -> Branding/Logo) page which will be displayed above the 'CurrentLogo' image after uploading a image, before saving it so that when a user uploads a image he can see whether it is correct or not(as shown in the below image). Please let us know your thoughts on this.
I like the preview idea.

Let me see if I can get a little previewlogo.

Thanks for the follow-up Keerthi!

-Nick
What size for the favicon?

Hi Nick,

Current favicon size is 1.4kb and the dimensions are 16*16 px.