Comments - Defect Number: 3343; Defect RecordID: 3339; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: Able to add an internal project through UI; Defect Description: On Tue, Mar 15, 2016 at 8:51 PM, Nick Callahan<nickc@logic2020.com>wrote: My responses inline.
Nick Callahan, IT Manager //Logic20/20//nickc@logic2020.com
t.206-576-0409//www.logic2020.com

From:Sai Keerthi Madireddy [mailto:keerthi.madireddy@ggktech.com]
Sent:Tuesday, March 15, 2016 8:00 AM
To:Nick Callahan <nickc@logic2020.com>; Kris Rushing Raines <KrisR@logic2020.com>
Cc:Logic 20/20 Core Team <logic2020@ggktech.com>
Subject:AddInternal/External field in Project Detail Page

Hi Nick,

Please find the following questions/assumptions on "Addinga fieldInternal/External in Project Detail Page" task and let us know your response.
@In PM, only people with role as 'System Administrator',@'Business unit Manager', 'Client Director', 'Practice Area Manager', 'Sales Persons', 'Senior Leadership', 'Operations' areabletoadda Project. Do you want us to follow the same criteria toaddainternalProject also?
That is fine.
In PM, we do not show Project Detail page forInternalProject. O
nce a project is added asinternalwe can not edit or view that project from UI
@to maintain consistency@
That is fine
@We show "Internal/External field" only inAddProject page , but not in Project Detail page. Please confirm.@
Since we will not have a Project Detail page (based on response to #2), that is fine.
@ForInternalProjects in PM, we only require Account, Business unit, Business Group, Practice Area fields, Project Name, Project status, Project Number.@Currently, we populate Account and Practice Area fields with "Logic20/20" and "Administration" values respectively.
That is fine.
As we are not using most of the fields toaddaninternalproject, we are thinking toadda new panel toaddanInternalproject in "Addproject" page as shown in the below images.

@
@@That is fine.
@
@Once user clicks on 'internal' tab in 'Addproject' page, user will be navigated to the above screen. please find the below notes for each field in the above screen.
Project Name:The name of theinternalproject to be added.
Project Number:It is a drop down list and is populated withInternalproject number series. When user clicks on save button the project number will be auto generated with the selected series from the drop down. Below are the project number series forinternalprojects which are going to be populated in the drop down.
 P91, P92, P93, P94, P95, P97.
Project Status:It is a drop down list populated with 'Active' or 'Internal' values(as theinternalprojects with these status are only get populated in 'Internal' section of time entry page).
Account, Business Unit, and Practice Area:These fields are auto populated with 'Logic20/20', 'Operations', and 'Administration' respectively.
@What is the difference between Active and Internal in this contextit seem that this could be prefilled with a status of Internal.
@
The project statuses 'Internal' and 'Active' are just statuses in Project status drop-down list. Project status 'Internal' does not point to theInternalprojects.
In PM, for all theinternalprojects we have status as "Active". But ininternalsection of the 'Time Entry' page we are populating all the projects with status as 'Active' or 'Internal', so we are thinking of populating the project status drop down with 'Active' and 'Internal' values.
Since we have all theinternalprojects with status as 'Active' we can auto populate Project status with 'Active'. Please let us know if we are wrong.@