Comments - Defect Number: 3350; Defect RecordID: 3346; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: Badge Date population issue; Defect Description: Hi Keerthi,
We identified anissuewithbadgedatefunctionality. The example is Tom Hun. We deactivated Tomsbadgeon 7/15/15. This populated organic breakdatesof 7/16/15  1/15/16. The organic enddatepopulated an 18 mos Period Start of 1/16/16.(Thisdateis not an actual startdatebut just an indicator of when they are available to be badged next.) We entered Tom on P000678 with abadgestart of 2/22/2016, however the 18 mos Period Start did change from 1/16/16. In this circumstance the project entry should override thedatepopulated by the organic break.

Let me know what questions you have.

Thanks,