Comments - Defect Number: 3353; Defect RecordID: 3349; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: Expenses record issue in Project Detail page; Defect Description: Keerthi,
Can you please take a look at this examplesee if we can come up with some sort of explanation as to what might be occurring.

Thanks!

-Nick


Hi Nick,
This is definitely a bug. All of theseexpenseitems were saved from7/1/2016 to 6/30/2017in P000966 for SMSGR. I even closed my browser and logged back into PM after setting them up to assure they were set to the right dates.

One day later and now theyve all shifted and are showing up in this fiscal. (See screenshot 1 below for theexpensedates, screenshot 2 for the summary view that clued us off to this).

Im going to manually enter these in again, save the project, and send you a screenshot that its updated. My biggest concern is that this bug will alter our financial projections to make the project look more profitable than it actually is.

Can we try a for a fix again? Both Tim and I have experienced this happening.

Thanks,
Preston