Comments - Defect Number: 3361; Defect RecordID: 3357; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: W2 Hourly Enhancements- Able to enter PTO hours for W2 Hourly salaried persons; Defect Description: Hi Ellen,
We have anenhancementrequest from Chad that we wanted to runbyyou. Currently,W2-Hourlyemployees can enter Sick/Safe Leave but not PTO.W2-Hourlyemployees do not appear on the PTO report.

Chad would like to add the PTO time entry option and add theW2-Hourlyemployees to the PTO report. What is the impact for you? Any other thoughts?

If we want to green light this request it could be a fairly quick win and could buy some more time while we gather requirements for the other items in the queue.

Thanks,

Kris Rushing Raines| Consulting Operations Lead |Logic20/20

Keerthi,

Can you please look at the below and let me know a ballpark level of effort?...that will help us to prioritize it against the rest of the queue. Let me know if you have questions on the description.

-Nick

Hi Nick,

It will take approximately 10 hours of work including changes in stored procedures, Handling data flow in UI and Testing.

Thank you