Comments - Defect Number: 3363; Defect RecordID: 3359; Server Address: localhost; Server Port: 99; Database Name: PM1; Defect Summary: New Project Detail url with Project Number; Defect Description: One more item.how much effort would it be to place theprojectnumber in theURL? For example, when I search for aprojectand then click on it,projectP000962 currently take me to:https://practice.logic2020.com/ProjectDetail.aspx?id=6658&returnTo=%7e%2fProjects.aspx
Wed want it something like:https://practice.logic2020.com/ProjectDetail.aspx?id=P000962

So other applications (or people) could actually create theURLbased on a pattern.

Can you generate an estimate for the effort to make that change?

Thanks!

Hi Nick,

In PM, we have many pages which have hyperlink toProjectDetailspage withURLhttps://practice.logic2020.com/ProjectDetail.aspx?id=6658. If we change the currentURL, we need to update code in all the pages which will take more effort(minimum of 24 hours). Hence, we are thinking to give access through both theurls
1.https://practice.logic2020.com/ProjectDetail.aspx?id=6658.
2.https://practice.logic2020.com/ProjectDetail.aspx?ProjectNumber=P000962(changing the query string argument name from 'id' to 'ProjectNumber')

So that you can access theprojectdetailspage with bothurlseither with ProjectNumber or Id, this way all hyperlinks in the application will still use the Id and if you need to accessprojectdetailsstill you can access with the ProjectNumberurlfrom browser.

Please let us know your response.

That will be fine and makes good sense.we in essence setup an aliasURL.

Let me know how much effort and if the below will be THE targetURLformat.so I can have other applications code to the below pattern immediately.

-Nick

It will take 4 hours of effort and please let us know the priority of the task.